{{cookiecutter.app_name}}-appsrv/requirements.txt
pyramid
pyramid-jinja2
pyramid-sqlalchemy
cornice
ujson
waitress